refactor: the wire's requests parse in request.rs
proto.rs sat at 399 of the 400 hard cap, so the next request on this
wire needed the split first. The Request enum, parse_request and the
request-side tests move whole to request.rs; proto.rs keeps the
one-line responses and theirs. One import line in run.rs, no behaviour.
